requirements.txt
requests>=2.31.0
json-stream>=1.3.0 
pdfplumber>=0.10.0
python-dateutil>=2.8.2
spacy>=3.7.2
nltk>=3.8.1 
pandas>=1.5.0
numpy>=1.23.0
scikit-learn>=1.2.0
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4
networkx>=3.0
openpyxl>=3.1.0
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
pydantic>=2.0.0 
python-dotenv>=1.0.0 
google-generativeai>=0.3.0
